producer/requirements.txt
requests
openai-whisper
pytube
pydub
Flask